+++ ide-common/src/main/java/org/digma/intellij/plugin/persistence/PersistenceService.kt
@@ -61,5 +61,9 @@ open class PersistenceService : PersistentStateComponent<PersistenceData> {
         state.isFirstTimePluginLoaded = true
     }
 
+    fun setNoInsightsYetNotificationPassed() {
+        state.noInsightsYetNotificationPassed = true
+    }
+
 
 }